Added disclaimer
Added "Disclaimer: This is not an official web site of the Hilliard
City School District. The Hilliard City School District does not
control and cannot guarantee the relevance, timeliness, or accuracy of
the information on this web site. Any views or opinions expressed
herein are solely those of the creators of this web site." to the footer